load cmt + add cmt ajax

+++ build/web/js/ajax.js
@@ -22,8 +22,9 @@ function handleAjaxMain(element) {
         selectSearch: select.value,
         inputSearch: input.value
     }
-    
-    if(element.closest('.RC-pagination a')) parameters.page = element.dataset.index;
+
+    if (element.closest('.RC-pagination a'))
+        parameters.page = element.dataset.index;
 
     $.ajax({
         url: "/Movie_Web/search-movie",
@@ -38,9 +39,8 @@ function handleAjaxMain(element) {
         }
     });
 }
-
-function handleInputSubmit(event, element){
-    if(event.key == 'Enter'){
+function handleInputSubmit(event, element) {
+    if (event.key == 'Enter') {
         event.preventDefault();
         handleAjaxMain(element);
     }
@@ -50,42 +50,42 @@ function handleInputSubmit(event, element){
 // ket hop voi ajax de tranh nhieu request nen moi lam the nay,
 // phan delete thi goi ajax
 // handle with shop-card
-function handleCheckAll(element){
+function handleCheckAll(element) {
     // let cardItems = $$('.shopping-card-item')
     let checkItems = _$_$('input[name="item-index"]')
 
-    for(let checkItem of checkItems){
+    for (let checkItem of checkItems) {
         checkItem.checked = element.checked;
     }
     handleMoneys()
 }
-function handleCheckItem(element){
+function handleCheckItem(element) {
     let checkAll = _$('#markAll')
     let checkItems = _$_$('input[name="item-index"]')
     handleMoneys()
-    for(let checkItem of checkItems){
-        if(!checkItem.checked) {
+    for (let checkItem of checkItems) {
+        if (!checkItem.checked) {
             checkAll.checked = false;
             return;
         }
     }
     checkAll.checked = true
 }
-function handleDelete(element){
+function handleDelete(element) {
     deleteTarget = element.closest('.shopping-card-item')
     deleteParent = deleteTarget.closest('.shopping-card-list')
     deleteParent.removeChild(deleteTarget)
     handleMoneys()
     handleCountItem()
 }
-function handleMoneys(){
+function handleMoneys() {
     let moneyPos = _$('.shopping-card-total span')
     let checkItems = _$_$('input[name="item-index"]')
     let stringMoneys = [], result
-    for(let checkItem of checkItems){
-        if(checkItem.checked){
+    for (let checkItem of checkItems) {
+        if (checkItem.checked) {
             let parentElement = checkItem.closest('.shopping-card-item')
-            if(parentElement.style.display != 'none'){
+            if (parentElement.style.display != 'none') {
                 let moneyPosTmp = parentElement.querySelector('.price-delete-wrapper p')
                 stringMoneys.push(moneyPosTmp.innerHTML)
             }
@@ -96,30 +96,30 @@ function handleMoneys(){
     }, 0)
     moneyPos.innerHTML = `$${result}`;
 }
-function handleCountItem(){
+function handleCountItem() {
     let countPos = _$('.SCL-heading span')
     let items = _$_$('.shopping-card-item')
-    let num=0;
-    for(let item of items){
-        if(item.style.display != 'none') num += 1;
+    let num = 0;
+    for (let item of items) {
+        if (item.style.display != 'none')
+            num += 1;
     }
     countPos.innerHTML = `${num} items`
 }
 
 //for search input with only js
-function handldeSearchCart(element){
+function handldeSearchCart(element) {
     let items = _$_$('.shopping-card-item');
     let searhValue = element.value.toUpperCase();
-    for(let item of items){
+    for (let item of items) {
         let name = item.querySelector('.card-film-info a').innerHTML.toUpperCase()
         let director = item.querySelector('.card-film-info p').innerHTML.toUpperCase()
         let dateCreated = item.querySelector('.card__film-created p').innerHTML.toUpperCase()
         let price = item.querySelector('.price-delete-wrapper p').innerHTML.toUpperCase()
         let isMatch = false;
-
-        if(name.indexOf(searhValue) > -1 || director.indexOf(searhValue) > -1 || dateCreated.indexOf(searhValue) > -1 || price.indexOf(searhValue) > -1)
+        if (name.indexOf(searhValue) > -1 || director.indexOf(searhValue) > -1 || dateCreated.indexOf(searhValue) > -1 || price.indexOf(searhValue) > -1)
             isMatch = true;
-        if(isMatch)
+        if (isMatch)
             item.style.display = ""
         else
             item.style.display = "none"
@@ -127,8 +127,7 @@ function handldeSearchCart(element){
     handleMoneys()
     handleCountItem()
 }
-
-function handleAjaxShopCart(element){
+function handleAjaxShopCart(element) {
     $.ajax({
         url: "/Movie_Web/shop-cart",
         type: "post", //send it through get method
@@ -138,13 +137,26 @@ function handleAjaxShopCart(element){
         success: function (data) {
 //            console.log(JSON.stringify(data));
             let dataRes = data
-            if(dataRes == "add") element.innerHTML = 'CANCEL CART'
-            if(dataRes == "delete") {
-                if(element.innerHTML == 'CANCEL CART'){
+            if (dataRes == "add") {
+                //for btn in movie detaiil
+                if (element.closest('.btn_add-to-cart')) {
+                    let textEle = element.querySelector('span')
+                    textEle.innerHTML = 'CANCEL CART'
+                } else
+                    element.innerHTML = 'CANCEL CART'
+            }
+            if (dataRes == "delete") {
+                if (element.innerHTML == 'CANCEL CART') {
                     element.innerHTML = 'ADD TO CART'
-                }
-                else {
-                    handleDelete(element)
+                } else {
+                    //for btn in movie detaiil
+                    if (element.closest('.btn_add-to-cart')) {
+                        let textEle = element.querySelector('span')
+                        textEle.innerHTML = 'ADD TO CART'
+                    }
+                    //for trah in shop-card
+                    else if (element.closest('.trash_delete-cart'))
+                        handleDelete(element)
                 }
             }
         },
@@ -152,4 +164,66 @@ function handleAjaxShopCart(element){
             //Do Something to handle error
         }
     });
+}
+
+
+//movie-detail============
+function handleAjaxLoadCmt(element) {
+    let numOfCmt = document.querySelectorAll('.comment-list .comment-item').length;
+    let filmId = element.dataset.id;
+    let parameters = {numOfCmt, filmId, action: 'load'}
+    
+    $.ajax({
+        url: "/Movie_Web/movie-detail",
+        type: "post", //send it through get method
+        data: parameters,
+        success: function (data) {
+            // khai bao bien kieu nay ~ JSON.stringify()
+            let dataString = JSON.stringify(data)
+            let dataLoadPos = document.querySelector('.comment__item-wrapper')
+            if(dataString.indexOf('no-more')>-1){
+//                console.log(dataString);
+                let newData = dataString.replace("|no-more", "")
+                dataLoadPos.innerHTML += JSON.parse(newData)
+                
+                let elementPar = element.closest('.comment-list')
+                elementPar.removeChild(element)
+            }
+            else{
+                dataLoadPos.innerHTML += data
+            }
+        },
+        error: function (xhr) {
+            //Do Something to handle error
+        }
+    });
+}
+function handleAjaxAddCmt(e, element) {
+    if(e.key == 'Enter'){
+        let inputCmt = element.value;
+        let uiId = element.dataset.uiid
+        let mId = element.dataset.mid
+        let parameters = {inputCmt, uiId, mId, action:'add'}
+        
+//        console.log(parameters)
+        $.ajax({
+        url: "/Movie_Web/movie-detail",
+        type: "post", //send it through get method
+        data: parameters,
+        success: function (data) {
+            let dataLoadPos = document.querySelector('.comment__item-wrapper')
+            dataLoadPos.innerHTML = data + dataLoadPos.innerHTML
+            element.value=''
+            CountCmtAfterAdd()
+        },
+        error: function (xhr) {
+            //Do Something to handle error
+        }
+    });
+    }
+}
+function CountCmtAfterAdd(){
+    let countPos = document.querySelector('.MD__main-comment h1 span')
+    let preCount = countPos.innerHTML
+    countPos.innerHTML = Number(preCount) + 1
 }
